+++ lib/src/boxes/boxes.dart
@@ -1,4 +1,11 @@
-export 'hivez_box.dart';
-export 'hivez_box_lazy.dart';
-export 'hivez_isolated.dart';
-export 'hivez_isolated_lazy.dart';
+library;
+
+import 'package:hive_ce/hive.dart';
+import 'package:meta/meta.dart';
+import 'package:synchronized/synchronized.dart';
+
+part 'base_box.dart';
+part 'hivez_box.dart';
+part 'hivez_box_lazy.dart';
+part 'hivez_isolated.dart';
+part 'hivez_isolated_lazy.dart';
