+++ lib/src/boxes/hivez_box.dart
@@ -1,14 +1,11 @@
-import 'package:hive_ce/hive.dart';
-import 'package:meta/meta.dart';
-
-import '../core/core.dart';
+part of 'boxes.dart';
 
 class HivezBox<K, T> extends AbstractHivezBox<K, T, Box<T>> {
   @override
   bool get isLazy => false;
 
   HivezBox(
-    super.boxName, {
+    super.name, {
     super.encryptionCipher,
     super.crashRecovery,
     super.path,
@@ -17,40 +14,26 @@ class HivezBox<K, T> extends AbstractHivezBox<K, T, Box<T>> {
   });
 
   @override
-  @protected
-  Box<T> hiveGetBox() => Hive.box<T>(boxName);
-
-  @override
-  @protected
-  Future<Box<T>> hiveOpenBox() async => await Hive.openBox<T>(
-        boxName,
-        encryptionCipher: encryptionCipher,
-        crashRecovery: crashRecovery,
-        path: path,
-        collection: collection,
-      );
-
-  @override
   Future<T?> get(K key, {T? defaultValue}) async {
-    return synchronizedRead(
-        () => Future.value(box.get(key, defaultValue: defaultValue)));
+    return _synchronizedRead(
+        () => Future.value(hiveBox.get(key, defaultValue: defaultValue)));
   }
 
   @override
   Future<Iterable<T>> getAllValues() async {
-    return synchronizedRead(() => Future.value(box.values));
+    return _synchronizedRead(() => Future.value(hiveBox.values));
   }
 
   @override
   Future<T?> valueAt(int index) async {
-    return synchronizedRead(() => Future.value(box.getAt(index)));
+    return _synchronizedRead(() => Future.value(hiveBox.getAt(index)));
   }
 
   @override
   Future<T?> firstWhereOrNull(bool Function(T) condition) async {
-    return synchronizedRead(() async {
-      for (final key in box.keys) {
-        final value = box.get(key);
+    return _synchronizedRead(() async {
+      for (final key in hiveBox.keys) {
+        final value = hiveBox.get(key);
         if (value != null && condition(value)) return value;
       }
       return null;
@@ -58,6 +41,20 @@ class HivezBox<K, T> extends AbstractHivezBox<K, T, Box<T>> {
   }
 
   Future<Map<K, T>> toMap() async {
-    return synchronizedRead(() => Future.value(box.toMap().cast<K, T>()));
+    return _synchronizedRead(() => Future.value(hiveBox.toMap().cast<K, T>()));
   }
+
+  @override
+  Future<Box<T>> _ceateBoxInHive() {
+    return Hive.openBox<T>(
+      name,
+      encryptionCipher: encryptionCipher,
+      crashRecovery: crashRecovery,
+      path: path,
+      collection: collection,
+    );
+  }
+
+  @override
+  Box<T> _getBoxFromHive() => Hive.box<T>(name);
 }
