+++ lib/src/boxes/hivez_box_lazy.dart
@@ -1,14 +1,11 @@
-import 'package:hive_ce/hive.dart';
-import 'package:meta/meta.dart';
-
-import '../core/core.dart';
+part of 'boxes.dart';
 
 class HivezLazyBox<K, T> extends AbstractHivezBox<K, T, LazyBox<T>> {
   @override
   bool get isLazy => true;
 
   HivezLazyBox(
-    super.boxName, {
+    super.name, {
     super.encryptionCipher,
     super.crashRecovery,
     super.path,
@@ -17,13 +14,8 @@ class HivezLazyBox<K, T> extends AbstractHivezBox<K, T, LazyBox<T>> {
   });
 
   @override
-  @protected
-  LazyBox<T> hiveGetBox() => Hive.lazyBox<T>(boxName);
-
-  @override
-  @protected
-  Future<LazyBox<T>> hiveOpenBox() async => await Hive.openLazyBox<T>(
-        boxName,
+  Future<LazyBox<T>> _ceateBoxInHive() => Hive.openLazyBox<T>(
+        name,
         encryptionCipher: encryptionCipher,
         crashRecovery: crashRecovery,
         path: path,
@@ -31,18 +23,21 @@ class HivezLazyBox<K, T> extends AbstractHivezBox<K, T, LazyBox<T>> {
       );
 
   @override
+  LazyBox<T> _getBoxFromHive() => Hive.lazyBox<T>(name);
+
+  @override
   Future<T?> get(K key, {T? defaultValue}) async {
-    return synchronizedRead(
-        () => Future.value(box.get(key, defaultValue: defaultValue)));
+    return _synchronizedRead(
+        () => Future.value(hiveBox.get(key, defaultValue: defaultValue)));
   }
 
   @override
   Future<Iterable<T>> getAllValues() async {
-    return synchronizedRead(() async {
-      final keys = box.keys.cast<K>();
+    return _synchronizedRead(() async {
+      final keys = hiveBox.keys.cast<K>();
       final List<T> values = [];
       for (final key in keys) {
-        final value = await box.get(key);
+        final value = await hiveBox.get(key);
         if (value != null) values.add(value);
       }
       return values;
@@ -51,14 +46,14 @@ class HivezLazyBox<K, T> extends AbstractHivezBox<K, T, LazyBox<T>> {
 
   @override
   Future<T?> valueAt(int index) async {
-    return synchronizedRead(() => box.getAt(index));
+    return _synchronizedRead(() => hiveBox.getAt(index));
   }
 
   @override
   Future<T?> firstWhereOrNull(bool Function(T) condition) async {
-    return synchronizedRead(() async {
-      for (final key in box.keys) {
-        final value = await box.get(key);
+    return _synchronizedRead(() async {
+      for (final key in hiveBox.keys) {
+        final value = await hiveBox.get(key);
         if (value != null && condition(value)) return value;
       }
       return null;
