--- lib/src/core/invalid_type_exception.dart
@@ -1,6 +1,0 @@
-class HivezBoxInvalidTypeException implements Exception {
-  final String message;
-  HivezBoxInvalidTypeException(this.message);
-  @override
-  String toString() => 'HivezBoxInvalidTypeException: $message';
-}
