--- lib/src/core/log_handler.dart
@@ -1,1 +1,0 @@
-typedef LogHandler = void Function(String message);
