--- lib/src/core/service_init_exception.dart
@@ -1,7 +1,0 @@
-class HiveServiceInitException implements Exception {
-  final String message;
-  HiveServiceInitException(this.message);
-
-  @override
-  String toString() => 'HiveServiceInitException: $message';
-}
