+++ lib/src/extensions/extensions.dart
@@ -1,1 +1,3 @@
+export 'backup_compressed.dart';
 export 'backup_json.dart';
+export 'search.dart';
