+++ lib/src/src.dart
@@ -1,3 +1,2 @@
-export 'core/core.dart';
 export 'boxes/boxes.dart';
 export 'extensions/extensions.dart';
